struct renaming.

+++ src/alias.h
@@ -27,9 +27,9 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 struct alias_s {
-  unsigned line_no;                     // line in conf file defined on
-  size_t argc;                          // number of arguments + 1
-  char const** argv;                    // argv[0] = alias name
+  unsigned      line_no;                // line in conf file defined on
+  unsigned      argc;                   // number of arguments + 1
+  char const  **argv;                   // argv[0] = alias name
 };
 typedef struct alias_s alias_t;
